Move recorded fill
